Prettify

+++ browser/patch-worker.js
@@ -23,7 +23,7 @@ limitations under the License.
 // Note: This script does not magically circumvent the Same origin policy.
 
 (function () {
-    'use strict';
+    "use strict";
     var Worker_ = window.Worker;
     var URL = window.URL || window.webkitURL;
     // Create dummy worker for the following purposes:
@@ -34,15 +34,16 @@ limitations under the License.
     // some of the used Chrome APIs are only supported as of Chrome 20,
     //  I don't bother adding a BlobBuilder fallback.
     var dummyWorker = new Worker_(
-        URL.createObjectURL(new Blob([], { type: 'text/javascript' })));
+        URL.createObjectURL(new Blob([], { type: "text/javascript" })),
+    );
     window.Worker = function Worker(scriptURL) {
         if (arguments.length === 0) {
-            throw new TypeError('Not enough arguments');
+            throw new TypeError("Not enough arguments");
         }
         try {
             return new Worker_(scriptURL);
         } catch (e) {
-            if (e.code === 18/*DOMException.SECURITY_ERR*/) {
+            if (e.code === 18 /*DOMException.SECURITY_ERR*/) {
                 return new WorkerXHR(scriptURL);
             } else {
                 throw e;
@@ -68,13 +69,13 @@ limitations under the License.
     function WorkerXHR(scriptURL) {
         var worker = this;
         var x = new XMLHttpRequest();
-        x.responseType = 'blob';
+        x.responseType = "blob";
         x.onload = function () {
             // http://stackoverflow.com/a/10372280/938089
             var workerURL = URL.createObjectURL(x.response);
             bindWorker(worker, workerURL);
         };
-        x.open('GET', scriptURL);
+        x.open("GET", scriptURL);
         x.send();
         worker._replayQueue = [];
         worker._messageQueue = [];
@@ -84,13 +85,12 @@ limitations under the License.
         terminate: function () {
             if (!this._terminated) {
                 this._terminated = true;
-                if (this.Worker)
-                    this.Worker.terminate();
+                if (this.Worker) this.Worker.terminate();
             }
         },
         postMessage: function (message, transfer) {
             if (!(this instanceof WorkerXHR))
-                throw new TypeError('Illegal invocation');
+                throw new TypeError("Illegal invocation");
             if (this.Worker) {
                 this.Worker.postMessage.apply(this.Worker, arguments);
             } else {
@@ -99,37 +99,42 @@ limitations under the License.
                 // Alright, push the valid message to the queue.
                 this._messageQueue.push(arguments);
             }
-        }
+        },
     };
     // Implement the EventTarget interface
-    [
-        'addEventListener',
-        'removeEventListener',
-        'dispatchEvent'
-    ].forEach(function (method) {
-        WorkerXHR.prototype[method] = function () {
-            if (!(this instanceof WorkerXHR)) {
-                throw new TypeError('Illegal invocation');
-            }
-            if (this.Worker) {
-                this.Worker[method].apply(this.Worker, arguments);
-            } else {
-                this._replayQueue.push({ method: method, arguments: arguments });
-            }
-        };
-    });
+    ["addEventListener", "removeEventListener", "dispatchEvent"].forEach(
+        function (method) {
+            WorkerXHR.prototype[method] = function () {
+                if (!(this instanceof WorkerXHR)) {
+                    throw new TypeError("Illegal invocation");
+                }
+                if (this.Worker) {
+                    this.Worker[method].apply(this.Worker, arguments);
+                } else {
+                    this._replayQueue.push({
+                        method: method,
+                        arguments: arguments,
+                    });
+                }
+            };
+        },
+    );
     Object.defineProperties(WorkerXHR.prototype, {
         onmessage: {
-            get: function () { return this._onmessage || null; },
+            get: function () {
+                return this._onmessage || null;
+            },
             set: function (func) {
-                this._onmessage = typeof func === 'function' ? func : null;
-            }
+                this._onmessage = typeof func === "function" ? func : null;
+            },
         },
         onerror: {
-            get: function () { return this._onerror || null; },
+            get: function () {
+                return this._onerror || null;
+            },
             set: function (func) {
-                this._onerror = typeof func === 'function' ? func : null;
-            }
-        }
+                this._onerror = typeof func === "function" ? func : null;
+            },
+        },
     });
 })();
